Fix test routes for email templates
For email notifications of posts, comments and users, the test routes
were putting the ID of the document they referred to in a field with the
wrong name (_id instead of documentId), causing that field to be
missing. This meant that when you used the test routes, the item you
selected in the URL would be ignored, and you'd get whatever a mongo
findOne() happened to return (ie, the first post in the database).